HAI-1359 Extract re-usable components from hanke form summary page and implement basic information and contacts tabs in hanke view
