+++ src/domain/hanke/edit/utils.ts
@@ -2,6 +2,7 @@ import { Feature } from 'ol';
 import Polygon from 'ol/geom/Polygon';
 import { Polygon as GeoJSONPolygon } from 'geojson';
 import { max, min } from 'date-fns';
+import { getArea } from 'ol/sphere';
 import { HankeDataDraft, HankeContact, HankeMuuTaho, HankeAlue } from '../../types/hanke';
 import { FORMFIELD, HankeDataFormState } from './types';
 import { formatFeaturesToHankeGeoJSON } from '../../map/utils';
@@ -82,3 +83,23 @@ export const convertHankeDataToFormState = (
   toteuttajat: hankeData?.toteuttajat ? hankeData.toteuttajat : [],
   muut: hankeData?.muut ? hankeData.muut : [],
 });
+
+/**
+ * Calculate total surface area of all hanke areas
+ */
+export function calculateTotalSurfaceArea(areas?: HankeAlue[]) {
+  try {
+    const areasTotalSurfaceArea = areas?.reduce((surfaceArea, currArea) => {
+      const feature = new Feature(
+        new Polygon(currArea.geometriat?.featureCollection.features[0]?.geometry.coordinates)
+      );
+      const geom = feature.getGeometry();
+      const currAreaSurface = geom && Math.round(getArea(geom));
+      return currAreaSurface ? surfaceArea + currAreaSurface : surfaceArea;
+    }, 0);
+
+    return areasTotalSurfaceArea;
+  } catch (error) {
+    return null;
+  }
+}
