+++ src/domain/hanke/hankeView/HankeView.tsx
@@ -16,6 +16,14 @@ import Container from '../../../common/components/container/Container';
 import Text from '../../../common/components/text/Text';
 import { HankeDataDraft } from '../../types/hanke';
 import styles from './HankeView.module.scss';
+import BasicInformationSummary from '../edit/components/BasicInformationSummary';
+import {
+  FormSummarySection,
+  SectionItemContent,
+  SectionItemTitle,
+} from '../../forms/components/FormSummarySection';
+import { calculateTotalSurfaceArea } from '../edit/utils';
+import ContactsSummary from '../edit/components/ContactsSummary';
 
 type Props = {
   hankeData?: HankeDataDraft;
@@ -25,6 +33,14 @@ type Props = {
 
 const HankeView: React.FC<Props> = ({ hankeData, onEditHanke, onDeleteHanke }) => {
   const { t } = useTranslation();
+
+  if (!hankeData) {
+    return null;
+  }
+
+  const areasTotalSurfaceArea = calculateTotalSurfaceArea(hankeData.alueet);
+
+  const { omistajat, rakennuttajat, toteuttajat, muut } = hankeData;
 
   return (
     <article className={styles.hankeViewContainer}>
@@ -80,10 +96,44 @@ const HankeView: React.FC<Props> = ({ hankeData, onEditHanke, onDeleteHanke }) =
                 <Tab>{t('hankePortfolio:tabit:yhteystiedot')}</Tab>
                 <Tab>{t('hankePortfolio:tabit:hakemukset')}</Tab>
               </TabList>
-              <TabPanel>Perustiedot</TabPanel>
+              <TabPanel>
+                <BasicInformationSummary formData={hankeData}>
+                  <SectionItemTitle>Hanke-alueiden kokonaispinta-ala</SectionItemTitle>
+                  <SectionItemContent>
+                    {areasTotalSurfaceArea && <p>{areasTotalSurfaceArea} m²</p>}
+                  </SectionItemContent>
+                </BasicInformationSummary>
+              </TabPanel>
               <TabPanel>Alueet</TabPanel>
               <TabPanel>Haittojen hallinta</TabPanel>
-              <TabPanel>Yhteystiedot</TabPanel>
+              <TabPanel>
+                <FormSummarySection>
+                  {omistajat && omistajat.length > 0 && (
+                    <ContactsSummary
+                      contacts={omistajat}
+                      title={t('form:yhteystiedot:titles:omistaja')}
+                    />
+                  )}
+                  {rakennuttajat && rakennuttajat.length > 0 && (
+                    <ContactsSummary
+                      contacts={rakennuttajat}
+                      title={t('form:yhteystiedot:titles:rakennuttajatPlural')}
+                    />
+                  )}
+                  {toteuttajat && toteuttajat?.length > 0 && (
+                    <ContactsSummary
+                      contacts={toteuttajat}
+                      title={t('form:yhteystiedot:titles:toteuttajatPlural')}
+                    />
+                  )}
+                  {muut && muut?.length > 0 && (
+                    <ContactsSummary
+                      contacts={muut}
+                      title={t('form:yhteystiedot:titles:muutPlural')}
+                    />
+                  )}
+                </FormSummarySection>
+              </TabPanel>
               <TabPanel>Hakemukset</TabPanel>
             </Tabs>
           </div>
